+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 
-- Feat: Attach Isolate name to thread context (#847)
-- Feat: Add Android thread to platform stacktraces (#853)
+* Feat: Attach Isolate name to thread context (#847)
+* Feat: Add Android thread to platform stacktraces (#853)
+* Fix: Rename auto initialize property (#857)
 
 ## 6.6.0-alpha.2