+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Feat: Add Android thread to platform stacktraces (#853)
 * Fix: Rename auto initialize property (#857)
+* Feat: Sync Scope to Native (#858)
 
 ## 6.6.0-alpha.2