+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Feat: Add Android thread to platform stacktraces (#853)
 * Fix: Rename auto initialize property (#857)
-* Bump: Sentry-Android to 6.0.0-beta.4
+* Bump: Sentry-Android to 6.0.0-beta.4 (#871)
 
 ## 6.6.0-alpha.2